Rename GUI Files
Rename GuiController to GuiView and GuiFunctions to GuiController to better match MVC
